+++ src/components/Footer/Footer.css
@@ -3,7 +3,7 @@
   background-color: #333;
   color: #fff;
   text-align: center;
-  padding: 10px 20px;
+  padding: 10px 0px;
   bottom: 0;
 }
 .footerpage-footer p {
