+++ src/components/Layout/BackGroundGrid.css
@@ -1,0 +1,8 @@
+.background-grid {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  background-color: #c4a1ff;
+  background-image: url("../../img/grid_temp2-18.svg");
+}
