+++ src/components/Layout/BackGroundGrid.js
@@ -1,0 +1,7 @@
+import "./BackGroundGrid.css";
+
+const BackGroundGrid = ({ children }) => {
+  return <div className="background-grid">{children}</div>;
+};
+
+export default BackGroundGrid;
