+++ src/components/Layout/CreatePageLayout.css
@@ -1,12 +1,3 @@
-.createpage-project-page-layout {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  background-color: #c4a1ff;
-  background-image: url("../../img/grid_temp2-18.svg");
-}
-
 .createpage-sub-nav {
   position: fixed;
   width: 100%;
@@ -21,13 +12,12 @@
   display: flex;
   flex-direction: column;
   width: 100%;
-  max-width: 800px;
   background-color: white;
-  padding: 20px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  margin-top: 120px; /* 서브 네비게이션 바 아래에 위치하도록 설정 */
+  padding: 20px 12% 20px 14%;
+  box-sizing: border-box;
+  border: 2px black solid;
+  border-top: none;
 }
-
 .createpage-create-project-form {
   display: flex;
   flex-direction: column;
