+++ src/components/Layout/PageLayout.css
@@ -1,0 +1,38 @@
+.page-layout {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 65.12%;
+  margin-left: 17.05%;
+  margin-right: 15.83%;
+  margin-bottom: 195px;
+  margin-top: 77px; /* 서브 네비게이션 바 아래에 위치하도록 설정 */
+  box-shadow: 5px 5px 0px 0px rgba(0, 0, 0, 1);
+}
+.three-color {
+  width: 100%;
+  background-color: white;
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  border: 2px black solid;
+  box-sizing: border-box;
+  padding: 3px;
+}
+.three-color-circle {
+  width: 9px;
+  height: 9px;
+  background-color: #000;
+  margin-right: 1%;
+  border-radius: 50%;
+  border: 1.5px black solid;
+}
+.three-green {
+  background-color: #34b2a7;
+}
+.three-yellow {
+  background-color: #ffd43a;
+}
+.three-red {
+  background-color: #ff8473;
+}
