+++ src/views/CreateProjectPage/CreateWithProjectPage.js
@@ -1,4 +1,4 @@
-import ProjectPageLayout from "../../components/Layout/ProjectPageLayout";
+import ProjectPageLayout from "../../components/Layout/CreatePageLayout";
 
 const CreateWithProjectPage = () => {
   return (
