+++ src/views/MainPage/MainPage.css
@@ -1,5 +1,4 @@
 html {
-  overflow: scroll;
   height: 100%;
 }
